working on user prof functionality
