+++ app/components/Navbar.jsx
@@ -25,7 +25,7 @@ const Navbar = ({current, list, changeCurrent, user, adminList, handleChange, in
 
         <div className="col-sm-5 col-md-5" >
         <Link to='/'><button className="btn btn-default">Home</button></Link>
-        {user ? <Link to='/Profile'><button className="btn btn-default">Profile</button></Link>: null}
+        {user ? <Link to="/Profile" ><button className="btn btn-default">Profile</button></Link>: null}
         {user ?(  user.isAdmin ? adminList.map((item)=> (
           <Link key ={item.id} to={item.name}>
               <button className="btn btn-default" onClick={() => changeCurrent(item)}> {item.name}</button>
