+++ app/containers/UserContainer.jsx
@@ -1,11 +1,10 @@
-import {retrieveUserInfo} from '../action-creators/user';
+import {retrieveUserAddress} from '../action-creators/address';
 import {connect} from 'react-redux';
 import User from '../components/User';
 
 const mapStateToProps = (state) => {
   return {
-    info: state.user.info,
-    address: state.user.address,
+    address: state.address.address,
     auth: state.auth
   }
 }
@@ -13,7 +12,7 @@ const mapStateToProps = (state) => {
 const mapDispatchToProps = (dispatch) => {
   return {
     findUserInfo: function(userInfo) {
-      dispatch(retrieveUserInfo(userInfo))
+      dispatch(retrieveUserAddress(userInfo))
     }
   }
 }
