+++ app/reducers/index.jsx
@@ -4,7 +4,7 @@ const rootReducer = combineReducers({
   auth: require('./auth').default,
   category: require('./category').default,
   orders: require('./orders').default,
-  user: require('./user').default,
+  address: require('./address').default,
   navbar: require('./navbar').default,
   products: require('./products').default,
   currentProduct: require('./product').default,
